D1457509: Some update
