+++ .github/workflows/ci_cd.yml
@@ -1250,7 +1250,7 @@ jobs:
           TESTS_VERSION: ${{ env.TESTS_VERSION }}
         run: |
           .venv\Scripts\Activate.ps1
-          uv sync --frozen --group tests --extra all
+          uv sync --frozen --group tests --extra all-dotnet
 
       - name: Wait for master cache update
         uses: ./.github/actions/check-cache
