+++ .github/workflows/sync-release.yml
@@ -1254,7 +1254,7 @@ jobs:
       - name: Install pyaedt and tests dependencies
         run: |
           .venv\Scripts\Activate.ps1
-          uv sync --frozen --group tests --extra all
+          uv sync --frozen --group tests --extra all-dotnet
 
       - name: Remove Ansys processes (if any)
         shell: powershell
