+++ .github/workflows/update-testmondata-cache.yml
@@ -963,7 +963,7 @@ jobs:
           TESTS_VERSION: ${{ env.TESTS_VERSION }}
         run: |
           .venv\Scripts\Activate.ps1
-          uv sync --frozen --group tests --extra all
+          uv sync --frozen --group tests --extra all-dotnet
 
       - name: Restore testmondata cache
         uses: ./.github/actions/testmon-cache
