+++ src/ansys/aedt/core/application/design.py
@@ -1211,7 +1211,7 @@ class Design(AedtObjects, PyAedtBase):
         >>> oProject.SetActiveDesign
         >>> oProject.InsertDesign
         """
-        if settings.use_multi_desktop:  # pragma: no cover
+        if settings.use_multi_desktop and self._desktop_class.is_grpc_api:  # pragma: no cover
             self._desktop_class.grpc_plugin.recreate_application(True)
             if self._design_name:
                 self._odesign = self.oproject.SetActiveDesign(self._design_name)
@@ -1258,7 +1258,7 @@ class Design(AedtObjects, PyAedtBase):
         >>> oDesktop.SetActiveProject
         >>> oDesktop.NewProject
         """
-        if settings.use_multi_desktop:  # pragma: no cover
+        if settings.use_multi_desktop and self._desktop_class.is_grpc_api:  # pragma: no cover
             self._desktop_class.grpc_plugin.recreate_application(True)
         return self._oproject
