+++ src/ansys/aedt/core/desktop.py
@@ -1197,7 +1197,7 @@ class Desktop(PyAedtBase):
         >>> d = Desktop()
         >>> d.odesktop
         """
-        if settings.use_grpc_api:
+        if self.is_grpc_api:
             tries = 0
             while tries < 5:
                 try:
