+++ tests/system/filter_solutions/conftest.py
@@ -30,6 +30,7 @@ from ansys.aedt.core import settings
 from ansys.aedt.core.filtersolutions import DistributedDesign
 from ansys.aedt.core.filtersolutions import LumpedDesign
 from tests.conftest import DESKTOP_VERSION
+from tests.conftest import USE_GRPC
 
 # Ensure US English locale (with cross-platform fallbacks)
 for _loc in ("en_US.UTF-8", "English_United States.1252", "en_US"):
@@ -46,12 +47,12 @@ else:
 @pytest.fixture
 def lumped_design():
     """Fixture for creating a LumpedDesign object."""
-    settings.use_grpc_api = False
+    settings.use_grpc_api = USE_GRPC
     return LumpedDesign(DESKTOP_VERSION)
 
 
 @pytest.fixture
 def distributed_design():
     """Fixture for creating a DistributedDesign object."""
-    settings.use_grpc_api = False
+    settings.use_grpc_api = USE_GRPC
     return DistributedDesign(DESKTOP_VERSION)
