+++ tests/system/filter_solutions/test_desktop_types.py
@@ -29,8 +29,10 @@ import pytest
 from ansys.aedt.core import Circuit
 from ansys.aedt.core import Hfss
 from ansys.aedt.core import Hfss3dLayout
+from ansys.aedt.core.filtersolutions_core.export_to_aedt import PartLibraries
 from ansys.aedt.core.generic.settings import is_linux
 from tests.conftest import DESKTOP_VERSION
+from tests.system.filter_solutions.resources import read_resource_file
 
 ON_CI = os.getenv("ON_CI", "false").lower() == "true"
 
@@ -51,6 +53,16 @@ class TestClass:
         assert variables["C1"].si_value == pytest.approx(1.967e-12)
         assert variables["L2"].si_value == pytest.approx(1.288e-8)
         assert variables["C3"].si_value == pytest.approx(6.366e-12)
+        circuit.desktop_class.close_desktop()
+
+    def test_import_tuned_variables(self, lumped_design):
+        lumped_design.export_to_aedt.simulate_after_export_enabled = True
+        lumped_design.export_to_aedt.optimize_after_export_enabled = True
+        lumped_design.export_to_aedt.part_libraries = PartLibraries.LUMPED
+        circuit = lumped_design.export_to_aedt.export_design()
+        assert lumped_design.export_to_aedt.import_tuned_variables().splitlines() == read_resource_file(
+            "imported_netlist.ckt", "Lumped"
+        )
         circuit.desktop_class.close_desktop()
 
     def test_distributed_circuit_exported_desktop(self, distributed_design):
