+++ tests/unit/test_desktop.py
@@ -102,6 +102,7 @@ def test_get_available_toolkits() -> None:
 def test_desktop_odesktop_retries(mock_settings, mock_sleep, mock_desktop) -> None:
     """Test Desktop.odesktop property retries to get the odesktop object."""
     desktop = Desktop()
+    desktop.is_grpc_api = True
     desktop.grpc_plugin = MagicMock()
     aedt_app = MagicMock()
     mock_odesktop = PropertyMock(name="oui", side_effect=[Exception("Failure"), aedt_app])
